# api-docs.yaml
openapi: 3.0.3
info:
  title: Geocoding API
  description: |
    A fast and comprehensive REST API for US ZIP code geocoding, Ohio address lookup, and county boundary data.
    
    ## 🔐 Authentication Required
    
    Most endpoints require an API key. To get started:
    
    1. **Register**: Visit the web interface and create an account
    2. **Get API Key**: Create an API key in your dashboard
    3. **Authenticate**: Use either authentication method:
       - `Authorization: Bearer your-api-key` header, OR
       - `X-API-Key: your-api-key` header
    4. **Test**: Use the "Authorize" button below to set your API key for all requests
    
    ## ✨ Features
    - **Lightning Fast**: PostgreSQL + PostGIS backed with optimized spatial indexes
    - **Multi-Level Coverage**: ZIP codes, cities, addresses, and county boundaries
    - **Comprehensive Data**: Population, density, coordinates, and geographic boundaries
    - **Flexible Search**: Multiple search modes with proximity and filtering
    - **ZIP Code Fallback**: City-based ZIP code lookup for flexible address searching
    - **Rich Metadata**: Timezone, county weights, administrative data
    - **GeoJSON Support**: Ready-to-use geographic data for mapping applications
    - **Usage Tracking**: Monitor your API usage and billing
    
    ## 📊 Data Coverage
    
    ### **🏙️ City Data**
    - **31,000+ US Cities**: Nationwide city coverage with demographics
    - **ZIP Code Associations**: All ZIP codes for each city
...
servers:
...
paths:
...
  /cities:
    get:
      summary: Search US Cities
      description: |
        Search for US cities using various filters including name, state, county, population, and location-based queries.
        
        The endpoint supports fuzzy text matching on city names and can filter by:
        - City name (partial match)
        - State (by ID or name)
        - County name
        - Minimum population
        - Geographic proximity (latitude/longitude with radius)
        
        Results are ordered by ranking (major cities first) and population.
        
        **Use Case**: Use this endpoint as a fallback when ZIP code searches fail. For example,
        if searching for an address with an incorrect ZIP code, search for the city to get all
        valid ZIP codes for that city, then search addresses in each ZIP.
      operationId: searchCities
      security:
        - ApiKeyAuth: []
      tags:
        - Cities
      parameters:
        - name: query
          in: query
          required: false
          description: Free text search query for city name
          schema:
            type: string
            example: "Dayton"
        - name: city
          in: query
          required: false
          description: Filter by city name (supports fuzzy matching)
          schema:
            type: string
            example: "Dayton"
        - name: state
          in: query
          required: false
          description: Filter by state ID (e.g., "OH") or state name (e.g., "Ohio")
          schema:
            type: string
            example: "OH"
        - name: county
          in: query
          required: false
          description: Filter by county name
          schema:
            type: string
            example: "Montgomery"
        - name: lat
          in: query
          required: false
          description: Latitude for location-based search (requires lng and optionally radius)
          schema:
            type: number
            format: double
            example: 39.7589
        - name: lng
          in: query
          required: false
          description: Longitude for location-based search (requires lat and optionally radius)
          schema:
            type: number
            format: double
            example: -84.1916
        - name: radius
          in: query
          required: false
          description: Search radius in kilometers (requires lat and lng)
          schema:
            type: number
            format: double
            example: 50
        - name: min_population
          in: query
          required: false
          description: Minimum population filter
          schema:
            type: integer
            example: 100000
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (max 100)
          schema:
            type: integer
            default: 10
            maximum: 100
            example: 10
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Cities found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitySearchResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cities/{id}:
    get:
      summary: Get City by ID
      description: |
        Retrieve detailed information for a specific city by its unique ID.
        
        Returns comprehensive city data including coordinates, population, associated ZIP codes,
        and administrative information.
      operationId: getCityById
      security:
        - ApiKeyAuth: []
      tags:
        - Cities
      parameters:
        - name: id
          in: path
          required: true
          description: Unique city identifier
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: City found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '400':
          description: Invalid city ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Supports two methods:
        
        **Method 1 - X-API-Key Header (Recommended):**
        ```
        X-API-Key: your-api-key-here
        ```
        
        **Method 2 - Authorization Bearer:**
        ```
        Authorization: Bearer your-api-key-here
        ```
        
        Get your API key by:
        1. Register/login at the web interface
        2. Create an API key in your dashboard
        3. Use either authentication method above
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: false
        error:
          type: string
          description: Error message describing what went wrong
          example: "ZIP code not found"
...
    City:
      type: object
      description: US city information including location, demographics, and associated ZIP codes
      properties:
        id:
          type: integer
          format: int64
          description: Unique city identifier
          example: 12345
        city:
          type: string
          description: City name
          example: "Dayton"
        city_ascii:
          type: string
          description: ASCII-only version of city name
          example: "Dayton"
        state_id:
          type: string
          description: Two-letter state code
          example: "OH"
        state_name:
          type: string
          description: Full state name
          example: "Ohio"
        county_fips:
          type: string
          description: County FIPS code
          example: "39113"
        county_name:
          type: string
          description: County name
          example: "Montgomery"
        lat:
          type: number
          format: double
          description: Latitude coordinate
          example: 39.7589
        lng:
          type: number
          format: double
          description: Longitude coordinate
          example: -84.1916
        population:
          type: integer
          description: City population
          example: 140407
        density:
          type: number
          format: double
          description: Population density (people per sq km)
          example: 1234.56
        source:
          type: string
          description: Data source
          example: "shape"
        military:
          type: boolean
          description: Whether this is a military base
          example: false
        incorporated:
          type: boolean
          description: Whether the city is incorporated
          example: true
        timezone:
          type: string
          description: IANA timezone identifier
          example: "America/New_York"
        ranking:
          type: integer
          description: City ranking (1 = major city)
          example: 1
        zips:
          type: string
          description: Space-separated list of ZIP codes
          example: "45401 45402 45403 45404 45405 45406 45409 45410 45414 45415"
        external_id:
          type: string
          description: External identifier from source data
          example: "1840034016"
    CitySearchResponse:
      type: object
      description: Response for city search operations
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/City'
        count:
          type: integer
          description: Number of results in current response
          example: 1
        total:
          type: integer
          description: Total number of matching cities
          example: 1
        query:
          type: string
          description: Original search query
          example: "Dayton"
        filters:
          type: object
          description: Applied filters
          additionalProperties: true
          example:
            state: "OH"
        message:
          type: string
          description: Optional message
        error:
          type: string
          description: Error message if success is false
    CityResponse:
      type: object
      description: Response for single city lookup
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/City'
        count:
          type: integer
          example: 1
        total:
          type: integer
          example: 1
...